Simplified code structure, updated playwright.config file to use baseURL, removed webkit from browser capabilities, updated md files for better understanding. and this commit is a check point for restructure in error handling and simplification of typescript funcitons.
